+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Updated the _experimental_ UI extensions APIs to allow top-level `await` in the initialization script.
 
 - Force unset the auth state of existing realtime connections on user password, collection secret, etc. changes.
-    _This is not strictly necessery because the realtime connections has short-lived idle timeout by design but nonetheless it was implemented to minimize the abuse vectors._
+    _This is not strictly necessery because the realtime connections have short-lived idle timeout by design but nonetheless it was implemented to restrict the abuse vectors._
 
 
 ## v0.38.0